peek-backend/requirements.txt
grpcio
grpcio-tools
requests
hume
hume[stream]
asyncio
python-dotenv
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
mindsdb_sdk
pybase64
pyairtable